Fix issue #5: crash when adding or removing profiles or folders during syncing
This fixes the crash caused by iterator invalidation when adding or removing elements in QList. It replaces QList with std::list because QList is now an alias for QVector, and QLinkedList has been removed from the Qt framework.

However, there was another way to fix the issue: using pointers and manually allocating and deallocating objects. But I decided to reject this idea to avoid any future memory leaks.