Fix copyFile() not copying modified date when disk transfer rate is used.
If the maximum disk transfer rate is not set, QFile::copy is used, which has two implementations: a native one and a custom one.
It first tries to copy a file using the native, platform-dependent function. If that fails, it then tries to copy the file using
the custom implementation, which reads from the file and writes to a new file with a new name using a buffer of 4096 bytes.

If the custom implementation is used, it does not copy the modified date, and we need to account for that. However, if we copy read-only files,
we cannot change their modification date after copying. Therefore, we will need a workaround to achieve this, or avoid using QFile::copy altogether.