Fix profile settings removal in profileNameChanged()
This is basically a fix for #27 issue. It turns out that every time a profile gets synced, profileNameChanged() is called. This triggers the removal of the profile settings while the app is still running, and it only writes the correct profile settings on quitting.
If the app crashes for any reason, that means all profile settings are lost. Not good. Not only that, profileNameChanged() gets called when the main window is shown for the first time, which also leads to that.