Fix incorrect pause status in some cases
Sometimes it shows an incorrect pause state, when all profiles are paused or at least one of them is unpaused. Also, sometimes the manager has an incorrect internal pause state, so we need to set it manually.
The code is not that great, but leave it as it is for now.