Remove unneeded file paths during scanning sync folders
When we scan synchronization folders for files, we load the paths of all found files into memory for any future synchronization operations.
Wouldn't it be nice to unload some of them to save memory if we're sure that no changes were detected and we don't need them anymore?
As a result, this gives us a 20%-50% reduction in memory usage during synchronization in my situation.

HOWEVER, it will likely decrease the performance of scanning of files a little bit. I don't have the accurate performance data at this moment.
Also, it's worth mentioning that we can only do this with files that were already in the database before synchronization.